Fixed Annotations filename issue
There are two variables, fileName and fileNameEl. The latter being the
DOM element that has the id="filename". The other is the value inside of
that element. It was stored for convenience but they were confused and a
file saving bug was introduced. They have been correctly named, declared
and initialized to avoid future problems.