Remove site-internal feedback system.
The user-facing feedback interface is unchanged, but it just goes straight to e-mail instead of storing in the database. This allows us to remove the Feedback-Viewing admin page which was unused and buggy anyway. For now, the database is left intact in case we want to review existing feedback, but we'll probably drop the feedback table eventually.