Eliminate Future[Option[T]] types.
We were swallowing a lot of errors in database access and web service
code behind options. Now all errors that can be explicitly handled are
propagated to the top-level controllers. They may still be ignored at
that point, but it should not be much easier to start improving our
error checking.